paquo.qupath: move jvm related code into qupath submodule
